+++ example/lib/dialogs.dart
@@ -51,8 +51,8 @@ class _DialogExamplesPageState extends State<DialogExamplesPage> {
                   Expanded(
                     child: DropdownSearch<int>(
                       items: (f, cs) => List.generate(30, (i) => i + 1),
-                      dropdownDecoratorProps: DropDownDecoratorProps(
-                        dropdownSearchDecoration: InputDecoration(labelText: "Dialog with title", hintText: "Select an Int"),
+                      decoratorProps: DropDownDecoratorProps(
+                        decoration: InputDecoration(labelText: "Dialog with title", hintText: "Select an Int"),
                       ),
                       popupProps: PopupProps.dialog(
                         title: Container(
@@ -80,8 +80,8 @@ class _DialogExamplesPageState extends State<DialogExamplesPage> {
                   Expanded(
                     child: DropdownSearch<int>(
                       items: (f, cs) => [1, 2, 3, 4, 5, 6, 7],
-                      dropdownDecoratorProps: DropDownDecoratorProps(
-                        dropdownSearchDecoration: InputDecoration(
+                      decoratorProps: DropDownDecoratorProps(
+                        decoration: InputDecoration(
                           labelText: "Modal mode",
                           hintText: "Select an Int",
                           filled: true,
@@ -250,8 +250,8 @@ class _DialogExamplesPageState extends State<DialogExamplesPage> {
                         ),
                       ),
                       compareFn: (item, selectedItem) => item.id == selectedItem.id,
-                      dropdownDecoratorProps: DropDownDecoratorProps(
-                        dropdownSearchDecoration: InputDecoration(
+                      decoratorProps: DropDownDecoratorProps(
+                        decoration: InputDecoration(
                           labelText: 'Users *',
                           filled: true,
                           fillColor: Theme.of(context).inputDecorationTheme.fillColor,
@@ -270,8 +270,8 @@ class _DialogExamplesPageState extends State<DialogExamplesPage> {
                         showSearchBox: true,
                       ),
                       compareFn: (item, sItem) => item.id == sItem.id,
-                      dropdownDecoratorProps: DropDownDecoratorProps(
-                        dropdownSearchDecoration: InputDecoration(
+                      decoratorProps: DropDownDecoratorProps(
+                        decoration: InputDecoration(
                           labelText: 'User *',
                           filled: true,
                           fillColor: Theme.of(context).inputDecorationTheme.fillColor,
@@ -375,8 +375,8 @@ class _DropdownWithGlobalCheckBoxState extends State<DropdownWithGlobalCheckBox>
     return DropdownSearch<int>.multiSelection(
       key: _infiniteScrollDropDownKey,
       items: (f, ic) => _getData(f, ic),
-      dropdownDecoratorProps: DropDownDecoratorProps(
-        dropdownSearchDecoration: InputDecoration(
+      decoratorProps: DropDownDecoratorProps(
+        decoration: InputDecoration(
           border: OutlineInputBorder(),
           labelText: 'Infinite Scroll',
           contentPadding: EdgeInsets.fromLTRB(12, 12, 0, 0),