+++ lib/src/properties/suggested_item_props.dart
@@ -1,3 +1,4 @@
+import 'package:dropdown_search/src/properties/scroll_props.dart';
 import 'package:flutter/material.dart';
 
 import '../../dropdown_search.dart';
@@ -17,11 +18,14 @@ class SuggestedItemProps<T> {
 
   final ClickProps itemClickProps;
 
+  final ScrollProps scrollProps;
+
   const SuggestedItemProps({
     this.suggestedItemBuilder,
     this.suggestedItems,
     this.suggestedItemsAlignment = MainAxisAlignment.start,
     this.showSuggestedItems = false,
     this.itemClickProps = const ClickProps(),
+    this.scrollProps = const ScrollProps(scrollDirection: Axis.horizontal),
   });
 }